backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.3
openai==1.3.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
pydantic==2.5.0
pydantic-settings==2.1.0
prometheus-client==0.19.0
aiofiles==23.2.1
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
PyJWT==2.8.0